# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fiestaaa_back",
    "description": "Backend API for Fiestaaa",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.2.1"
  },
  "paths": {
...
    "/events/{event_id}/expenses": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "list_event_expenses",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items to return. When omitted with cursor, the default is 50.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned by the previous response.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event shared expenses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventExpenseView"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "operationId": "create_event_expense",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventExpensePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Expense created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventExpenseView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/expenses/summary": {
      "get": {
        "tags": [
          "events"
        ],
        "operationId": "get_event_expenses_summary",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shared expense summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventExpensesSummaryView"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Event not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}/expenses/{expense_id}": {
      "delete": {
        "tags": [
          "events"
        ],
        "operationId": "delete_event_expense",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "expense_id",
            "in": "path",
            "description": "Expense identifier",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "EventExpenseBalanceView": {
        "type": "object",
        "required": [
          "user_id",
          "paid_cents",
          "owed_cents",
          "balance_cents"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "balance_cents": {
            "type": "integer",
            "format": "int64"
          },
          "handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "owed_cents": {
            "type": "integer",
            "format": "int64"
          },
          "paid_cents": {
            "type": "integer",
            "format": "int64"
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EventExpenseParticipantView": {
        "type": "object",
        "required": [
          "user_id"
        ],
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EventExpensePayload": {
        "type": "object",
        "required": [
          "title",
          "amount_cents",
          "participant_user_ids"
        ],
        "properties": {
          "amount_cents": {
            "type": "integer",
            "format": "int64"
          },
          "expense_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "paid_by_user_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "participant_user_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "EventExpenseSettlementView": {
        "type": "object",
        "required": [
          "from_user_id",
          "to_user_id",
          "amount_cents"
        ],
        "properties": {
          "amount_cents": {
            "type": "integer",
            "format": "int64"
          },
          "from_handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "from_user_id": {
            "type": "integer",
            "format": "int64"
          },
          "to_handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "to_user_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EventExpenseView": {
        "type": "object",
        "required": [
          "expense_id",
          "event_id",
          "paid_by_user_id",
          "title",
          "amount_cents",
          "expense_date",
          "created_at",
          "participants"
        ],
        "properties": {
          "amount_cents": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "event_id": {
            "type": "integer",
            "format": "int64"
          },
          "expense_date": {
            "type": "string",
            "format": "date-time"
          },
          "expense_id": {
            "type": "integer",
            "format": "int64"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "paid_by_avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "paid_by_handle": {
            "type": [
              "string",
              "null"
            ]
          },
          "paid_by_user_id": {
            "type": "integer",
            "format": "int64"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventExpenseParticipantView"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
      "EventExpensesSummaryView": {
        "type": "object",
        "required": [
          "currency",
          "total_expenses_cents",
          "balances",
          "settlements"
        ],
        "properties": {
          "balances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventExpenseBalanceView"
            }
          },
          "currency": {
            "type": "string"
          },
          "settlements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventExpenseSettlementView"
            }
          },
          "total_expenses_cents": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}